+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 ## Change log
 ----------------------
 
+Version 2
+-------------
+
+ADDED:
+
+- new constructor in class SimpleFilenameFilter with only the file extension
+- new factory methods in class SimpleFilenameFilter for both constructors created
+- new factory methods for all classes that implements the FileFilter interface
+
+CHANGED:
+
+- update of jdk to version 11
+- update of gradle to new version 7.5-rc-2
+- update of gradle-plugin dependency of 'com.diffplug.spotless:spotless-plugin-gradle' in version 6.7.2
+
 Version 1.9
 -------------